third_party/blink/web_tests/tables/mozilla_expected_failures/dom/appendCells1-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: appendCellAt is not defined
